requirements.txt
google-auth==2.3.0
google-cloud-speech==2.5.1
Flask==2.1.0
gunicorn==20.1.0
flask-cors